Platform refactor
